feat: restore apple game
