+++ src/pages/games/AppleGame/appleGameType.ts
