+++ src/pages/games/AppleGame/components/ExpBarChart.tsx
