+++ src/pages/games/AppleGame/components/GameStart.tsx
