+++ src/pages/games/AppleGame/components/SnackGameHUD.tsx
