+++ src/pages/games/AppleGame/model/appleGame/apple.ts
