+++ src/pages/games/AppleGame/model/appleGame/goldenApple.ts
@@ -1,4 +1,4 @@
-import GoldenAppleImage from '@assets/images/golden_apple.png';
+import GoldenAppleImage from '@assets/images/golden_snack.png';
 
 import Apple, { ApplePropType } from './apple';
