+++ src/pages/games/AppleGame/model/common/click.ts
