+++ src/pages/games/AppleGame/model/common/drag.ts
