+++ src/pages/games/AppleGame/model/common/particle.ts
