+++ src/pages/games/AppleGame/model/snackGame/goldSnack.ts
