+++ src/pages/games/AppleGame/model/snackGame/plainSnack.ts
