+++ src/pages/games/AppleGame/model/snackGame/snack.ts
