+++ src/pages/games/AppleGame/model/snackGame/snackGame.ts
