+++ src/pages/games/AppleGame/model/snackGame/snackGameInf.ts
