+++ src/pages/games/AppleGame/model/snackGame/snackGameMobile.ts
